Implement experimental fq_codel qdisc under htb
Instead of relying on the inferior sfq qdisc and secondary htb qdisc for device-level sharing, replace the secondary htb qdisc with fq_codel.

This is an opt-in option in the GUI to use the default qdisc from the QoS page (if any option is available) or override it with fq_codel.

It's really more than replacing sfq with fq_codel. It's replacing htb+htb+sfq with htb+fq_codel.