webui: Add support for exact matches in IP filter fields
In the Tracked connections table, filtering on IP has been based on partial matching in order to support other subnets like clients connected to the OpenVPN servers.

We now support exact matching by adding a $ to the end of the IP text. This is added automatically when choosing a client from the Local IP dropdown list, or can be typed manually in the field (e.g. 192.168.50.11$).