webui: Add user rules to well-known rules dropdown to allow rule reordering
A long-standing deficiency in FlexQoS has been the inability to change the order of the rules in the iptables rule list, where order is important.

We now add the user's existing iptables rules to be added to the Well-known rules dropdown, so that a user may delete any or all of the existing iptables rules and add them back in any order they choose.
The rule being added via the dropdown must not already exist in the iptables rule list. You must delete it first in order to add it again (there is still a check for duplicate rules on every change).