webui: use CDN version of chart.min.js for consistency on all models
Due to the breaking changes with Chart.js v3.9.1 coming in 3006.102.1
and 3004.388.8, no longer rely on chart.min.js from the firmware.

Instead we will load chart.min.js v2.9.4 from a CDN source to avoid
different coding and different behavior on 386, 388 and 102 firmware
branches.

Since FlexQoS runs best on older 386 and some 388 models, and
Adaptive QoS is currently broken on the 3006.102 models, this is in
the best interest of the existing users.